refactor: split enterprise integration tests from other tests
We now need to activate the maven profile `enterprise-integration-tests` to execute enterprise tests.

https://gravitee.atlassian.net/browse/APIM-3447
